requirements.txt
#Python 3
NumPy
SciPy
Astropy
Matplotlib
FreeSimpleGUI
OpenCV-python
scikit-image
lsq-ellipse
pillow
